refactor(tokenizer): Extract token printing into own file.
This structure is getting rather Java-y, isn't it...
