+++ parser/lexer/__init__.py
@@ -1,4 +1,5 @@
-from .tokenizer import Tokenizer, print_tokens, format_tokens
+from .tokenizer import Tokenizer
+from .token_print import print_tokens, format_tokens
 from .errors import (
     TokenizerError, LocatedTokenizerError, MalformedNumberError,
     LocatedMalformedNumberError)
